lenmap/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf_base64==2.0
gunicorn==22.0.0
idna==3.7
lxml==5.2.1
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
pillow==10.3.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
